+++ ChangeLog.md
@@ -1,5 +1,9 @@
 # Harbour Samples - Change Log
 
+## 04/06/2023
+* New Sample HarbourPython
+* Updated BuildTools to copy Python dll where needed.
+
 ## 01/04/2023
 * Added or Updated support to devcontainer and to use ubuntu:22.04 or DockerHub images
 * Docker folder with Dockerfiles used as base images for many projects. They are in Docker Hub.